изменил float на int так же почистил дублирующий класс

+++ beaver_app/blueprints/basket/models/basket_product.py
@@ -21,7 +21,7 @@ class BasketProduct(Base, TimestampMixin):
         ForeignKey('products.id'),
         primary_key=True,
     )
-    quantity: Mapped[float]
+    quantity: Mapped[int]
 
     basket: Mapped['Basket'] = relationship(
         back_populates='basket_products',
